refactor: move fallback movie data to constants file
